Refactor codebase for cross-platform support (Phase 1)
Add Go build tags to separate Windows-specific code from platform-independent
code, enabling the project to compile on macOS and Linux.

- Split platform/debug.go into cross-platform log.go + Windows-only debug_windows.go
- Add //go:build windows to all Win32-dependent files (ui/*.go, main, autostart, desktop)
- Create non-Windows stubs (autostart_other.go, debug_other.go, desktop_other.go)
- Add main_other.go with informational message for unsupported platforms
- Verified: builds successfully on windows, linux, darwin (amd64+arm64)

Closes #19 Phase 1. See #19 for remaining phases (macOS/Linux native UI).
