+++ internal/platform/autostart_windows.go
@@ -1,3 +1,5 @@
+//go:build windows
+
 package platform
 
 import (
