+++ internal/platform/log.go
@@ -4,15 +4,15 @@ import (
 	"fmt"
 	"os"
 	"path/filepath"
-	"syscall"
 	"time"
-	"unsafe"
 )
 
 var logFile *os.File
 
 // InitLog creates a log file next to the executable for diagnostics.
-// This is critical because -H windowsgui hides all stdout/stderr output.
+// On Windows this is critical because -H windowsgui hides all stdout/stderr output.
+// TODO(cross-platform): On macOS/Linux, consider using ~/.local/share/claude-hud/
+// or XDG_DATA_HOME instead of writing next to the binary.
 func InitLog() {
 	exePath, _ := os.Executable()
 	logPath := filepath.Join(filepath.Dir(exePath), "claude-hud.log")
@@ -41,35 +41,4 @@ func CloseLog() {
 	if logFile != nil {
 		logFile.Close()
 	}
-}
-
-// DLL procs used by platform package only
-var (
-	user32Plat   = syscall.NewLazyDLL("user32.dll")
-	procMessageBox = user32Plat.NewProc("MessageBoxW")
-)
-
-// ShowMessageBox displays a Windows MessageBox for critical errors.
-// Use this sparingly - only for fatal startup errors.
-func ShowMessageBox(title, message string) {
-	titlePtr, _ := syscall.UTF16PtrFromString(title)
-	msgPtr, _ := syscall.UTF16PtrFromString(message)
-	procMessageBox.Call(
-		0,
-		uintptr(unsafe.Pointer(msgPtr)),
-		uintptr(unsafe.Pointer(titlePtr)),
-		0x00000010, // MB_ICONERROR
-	)
-}
-
-// ShowInfoBox displays an informational MessageBox.
-func ShowInfoBox(title, message string) {
-	titlePtr, _ := syscall.UTF16PtrFromString(title)
-	msgPtr, _ := syscall.UTF16PtrFromString(message)
-	procMessageBox.Call(
-		0,
-		uintptr(unsafe.Pointer(msgPtr)),
-		uintptr(unsafe.Pointer(titlePtr)),
-		0x00000040, // MB_ICONINFORMATION
-	)
 }
