+++ internal/ui/window.go
@@ -1,3 +1,5 @@
+//go:build windows
+
 package ui
 
 import (
